Removed the line where we delete the tk library.
This was preventing building tclkit after we already built tclkit-gui.
This was most likely deleting the dll link library but in current 8.6
we don't actually produce one and this line removes the Tk library we need
for linking. Possibly this is needed for earlier builds - so commented
until it is checked against 8.4.

Signed-off-by: Pat Thoyts <patthoyts@users.sourceforge.net>